Lift the Beatport browse/scrape surface out of web_server.py
37 routes, the scrape cache and its accessors, the enrichment job pair and
clean_beatport_text move wholesale to api/beatport.py (-2,190 lines). This is
the self-contained slice: homepage sections, genre pages, top-100s, chart
extraction, release scraping. The discovery/sync/charts routes stay behind with
the discovery machinery they are welded to - the same split the discovery lift
chose. Verified by a full url-map diff (1,152 routes before and after, byte-
identical on method+path) plus an in-app smoke of the moved endpoints; the
automation's cache-refresh dependency still sees the same cache object.
